backend/requirements.txt
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.26.3
supabase==2.3.1
Pillow==10.2.0
Flask==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0